remove tailwindcss
