+++ package.json
@@ -13,15 +13,13 @@
     "gray-matter": "^4.0.2",
     "highlight.js": "^10.6.0",
     "next": "10.0.7",
-    "postcss": "^8.2.7",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "remark": "^13.0.0",
     "remark-highlight.js": "^6.0.0",
     "remark-html": "^13.0.1",
     "remark-slug": "^6.0.0",
-    "remark-toc": "^7.2.0",
-    "tailwindcss": "^2.0.3"
+    "remark-toc": "^7.2.0"
   },
   "devDependencies": {
     "@types/node": "^14.14.31",
