+++ src/styles/global.css
@@ -1,5 +1,3 @@
-@tailwind components;
-@tailwind utilities;
 
 a {
     color:black;
