--- tailwind.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    purge: ['./src/pages/**/*.{js,ts,jsx,tsx}', './src/components/**/*.{js,ts,jsx,tsx}'],
-    theme: {
-      extend: {},
-    },
-    variants: {},
-    plugins: [],
-  }
